third_party/blink/web_tests/fast/canvas/multiple-readbacks-without-will-read-frequently-produces-warning-expected.txt

CONSOLE WARNING: Canvas2D: Multiple readback operations using getImageData are faster with the willReadFrequently attribute set to true. See: https://html.spec.whatwg.org/multipage/canvas.html#concept-canvas-will-read-frequently
Verify that reading back from a canvas2D multiple times without using willReadFrequently produces a console warning.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

First pixel of canvas with willReadFrequently set to true: 255,0,255,255
Second pixel of canvas with willReadFrequently set to true: 0,0,0,0
First pixel of canvas with willReadFrequently set to false: 255,0,255,255
Second pixel of canvas with willReadFrequently set to false: 0,0,0,0
First pixel of canvas with willReadFrequently not set: 255,0,255,255
Second pixel of canvas with willReadFrequently not set: 0,0,0,0
PASS successfullyParsed is true

TEST COMPLETE
